szczecin_scraper/requirements.txt
# Szczecin Business Scraper - Dependencies
# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0
webdriver-manager>=4.0.0

# Google APIs
google-api-python-client>=2.100.0
googlemaps>=4.10.0

# Data processing
pandas>=2.1.0
openpyxl>=3.1.0  # Excel support
python-docx>=1.0.0  # DOCX support

# Rate limiting & retries
ratelimit>=2.2.1
tenacity>=8.2.0

# Async support
aiohttp>=3.9.0
asyncio>=3.4.3

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0  # Progress bars
fake-useragent>=1.4.0

# Optional: for proxy rotation
# rotating-proxies>=0.1.0
